feat: link mutation to the database
