+++ src/entity/User.ts
@@ -1,32 +1,16 @@
-import {
-  Entity,
-  PrimaryGeneratedColumn,
-  Column,
-  Connection,
-  BaseEntity,
-} from "typeorm";
+import { Entity, PrimaryGeneratedColumn, Column } from "typeorm";
 
 @Entity()
-export class User extends BaseEntity {
+export class User {
   @PrimaryGeneratedColumn()
   id: number;
 
   @Column()
-  firstName: string;
+  name: string;
 
   @Column()
-  lastName: string;
+  email: string;
 
   @Column()
-  age: number;
-}
-
-export async function createUser(database: Connection) {
-  const user = new User();
-  user.firstName = "Joao";
-  user.lastName = "Silva";
-  user.age = 35;
-
-  const newUser = await database.manager.save(user);
-  console.log("User has been saved. user id is", newUser.id);
+  birthDate?: string;
 }
